Move `Chunks` into the `preduce_reducer_script` crate
