+++ preduce_reducer_script/src/lib.rs
@@ -664,6 +664,39 @@ pub fn run_ranges<R: RemoveRanges>() -> ! {
     run::<RemoveRangesReducer<R>>()
 }
 
+/// A `RemoveRanges` implementation that removes chunks of lines from the seed
+/// file.
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Chunks;
+
+impl RemoveRanges for Chunks {
+    fn remove_ranges(seed: PathBuf) -> io::Result<Vec<Range<u64>>> {
+        let mut seed = fs::File::open(seed)?;
+        let mut ranges = vec![];
+
+        const BUF_SIZE: usize = 1024 * 1024;
+        let mut buf: Vec<u8> = vec![0; BUF_SIZE];
+
+        let mut start_of_line = 0;
+        let mut current_index = 0;
+        let mut bytes_read;
+        while {
+            bytes_read = seed.read(&mut buf)?;
+            bytes_read > 0
+        } {
+            for b in &buf[0..bytes_read] {
+                current_index += 1;
+                if *b == b'\n' {
+                    ranges.push(start_of_line..current_index);
+                    start_of_line = current_index;
+                }
+            }
+        }
+
+        Ok(ranges)
+    }
+}
+
 /// A trait for defining a regex that we use to implement a reducer that tries
 /// removing the regex's matches' capture groups from the seed test case.
 ///
